requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url --output-file=requirements.txt requirements.in
#
alembic==1.4.2            # via apache-airflow
apache-airflow==1.10.10   # via -r requirements.in
apispec[yaml]==1.3.3      # via flask-appbuilder
argcomplete==1.11.1       # via apache-airflow
attrs==19.3.0             # via apache-airflow, cattrs, jsonschema
babel==2.8.0              # via flask-babel
cached-property==1.5.1    # via apache-airflow
cachetools==4.1.1         # via google-auth
cattrs==0.9.2             # via apache-airflow
certifi==2020.4.5.2       # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via flask, flask-appbuilder
colorama==0.4.3           # via flask-appbuilder
colorlog==4.0.2           # via apache-airflow
configparser==3.5.3       # via apache-airflow
croniter==0.3.32          # via apache-airflow
defusedxml==0.6.0         # via python3-openid
dill==0.3.1.1             # via apache-airflow
dnspython==1.16.0         # via email-validator
docutils==0.16            # via python-daemon
email-validator==1.1.1    # via flask-appbuilder
flask-admin==1.5.4        # via apache-airflow
flask-appbuilder==2.3.4   # via apache-airflow
flask-babel==1.0.0        # via flask-appbuilder
flask-caching==1.3.3      # via apache-airflow
flask-jwt-extended==3.24.1  # via flask-appbuilder
flask-login==0.4.1        # via apache-airflow, flask-appbuilder
flask-openid==1.2.5       # via flask-appbuilder
flask-sqlalchemy==2.4.3   # via flask-appbuilder
flask-swagger==0.2.13     # via apache-airflow
flask-wtf==0.14.3         # via apache-airflow, flask-appbuilder
flask==1.1.2              # via apache-airflow, flask-admin, flask-appbuilder, flask-babel, flask-caching, flask-jwt-extended, flask-login, flask-openid, flask-sqlalchemy, flask-swagger, flask-wtf
funcsigs==1.0.2           # via apache-airflow
future==0.18.2            # via apache-airflow
google-api-core==1.21.0   # via google-cloud-core
google-auth==1.19.1       # via -r requirements.in, google-api-core, google-cloud-storage
google-cloud-core==1.3.0  # via google-cloud-storage
google-cloud-storage==1.29.0  # via -r requirements.in
google-resumable-media==0.5.1  # via google-cloud-storage
googleapis-common-protos==1.52.0  # via google-api-core
graphviz==0.14            # via apache-airflow
gunicorn==19.10.0         # via apache-airflow
idna==2.9                 # via email-validator, requests
importlib-metadata==1.7.0  # via argcomplete, jsonschema
iso8601==0.1.12           # via apache-airflow
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.10.3            # via apache-airflow, flask, flask-babel
json-merge-patch==0.2     # via apache-airflow
jsonschema==3.2.0         # via apache-airflow, flask-appbuilder
lazy-object-proxy==1.5.0  # via apache-airflow
lockfile==0.12.2          # via python-daemon
mako==1.1.3               # via alembic
markdown==2.6.11          # via apache-airflow
markupsafe==1.1.1         # via jinja2, mako, wtforms
marshmallow-enum==1.5.1   # via flask-appbuilder
marshmallow-sqlalchemy==0.23.1  # via flask-appbuilder
marshmallow==2.21.0       # via flask-appbuilder, marshmallow-enum, marshmallow-sqlalchemy
natsort==7.0.1            # via croniter
numpy==1.18.5             # via pandas
pandas==0.25.3            # via apache-airflow
pendulum==1.4.4           # via apache-airflow
prison==0.1.3             # via flask-appbuilder
protobuf==3.12.2          # via google-api-core, googleapis-common-protos
psutil==5.7.0             # via apache-airflow
psycopg2-binary==2.8.5    # via -r requirements.in
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pygments==2.6.1           # via apache-airflow
pyjwt==1.7.1              # via flask-appbuilder, flask-jwt-extended
pyrsistent==0.16.0        # via jsonschema
python-daemon==2.1.2      # via apache-airflow
python-dateutil==2.8.1    # via alembic, apache-airflow, croniter, flask-appbuilder, pandas, pendulum
python-dotenv==0.13.0     # via -r requirements.in
python-editor==1.0.4      # via alembic
python3-openid==3.1.0     # via flask-openid
pytz==2020.1              # via babel, flask-babel, google-api-core, pandas, tzlocal
pytzdata==2019.3          # via pendulum
pyyaml==5.3.1             # via apispec, flask-swagger
requests==2.23.0          # via -r requirements.in, apache-airflow, google-api-core
rsa==4.6                  # via google-auth
setproctitle==1.1.10      # via apache-airflow
six==1.15.0               # via flask-jwt-extended, google-api-core, google-auth, google-resumable-media, jsonschema, prison, protobuf, pyrsistent, python-dateutil, sqlalchemy-utils, tenacity, thrift
sqlalchemy-jsonfield==0.9.0  # via apache-airflow
sqlalchemy-utils==0.36.6  # via flask-appbuilder
sqlalchemy==1.3.17        # via -r requirements.in, alembic, apache-airflow, flask-sqlalchemy, marshmallow-sqlalchemy, sqlalchemy-jsonfield, sqlalchemy-utils
tabulate==0.8.7           # via apache-airflow
tenacity==4.12.0          # via apache-airflow
termcolor==1.1.0          # via apache-airflow
text-unidecode==1.2       # via apache-airflow
thrift==0.13.0            # via apache-airflow
typing-extensions==3.7.4.2  # via apache-airflow
tzlocal==1.5.1            # via apache-airflow, pendulum
unicodecsv==0.14.1        # via apache-airflow
urllib3==1.25.9           # via requests
werkzeug==0.16.1          # via apache-airflow, flask, flask-caching, flask-jwt-extended
wtforms==2.3.1            # via flask-admin, flask-wtf
zipp==3.1.0               # via importlib-metadata
zope.deprecation==4.4.0   # via apache-airflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
